Final update

+++ KvalDarbsCore/Controllers/CompetitionsController.cs
@@ -17,11 +17,11 @@ namespace KvalDarbsCore.Controllers
         {
         }
 
-        // GET: Competitions
+        [HttpGet]
         public ActionResult Index(string sortOrder, string searchString)
         {
-            ViewData["NameSort"] = string.IsNullOrEmpty(sortOrder) ? "name_desc" : string.Empty;
-            ViewData["DateSort"] = sortOrder == "Date" ? "date_desc" : "Date";
+            ViewData["NameSort"] = sortOrder == "Name" ? "name_desc" : "Name";
+            ViewData["DateSort"] = string.IsNullOrEmpty(sortOrder) ? "date_desc" : "";
             ViewData["LocationSort"] = sortOrder == "Location" ? "location_desc" : "Location";
             ViewData["Sort"] = sortOrder;
 
@@ -40,8 +40,8 @@ namespace KvalDarbsCore.Controllers
                 case "name_desc":
                     competitions = competitions.OrderByDescending(s => s.Name);
                     break;
-                case "Date":
-                    competitions = competitions.OrderBy(s => s.Date);
+                case "Name":
+                    competitions = competitions.OrderBy(s => s.Name);
                     break;
                 case "date_desc":
                     competitions = competitions.OrderByDescending(s => s.Date);
@@ -53,7 +53,7 @@ namespace KvalDarbsCore.Controllers
                     competitions = competitions.OrderByDescending(s => s.Location);
                     break;
                 default:
-                    competitions = competitions.OrderBy(s => s.Name);
+                    competitions = competitions.OrderBy(s => s.Date);
                     break;
             }
 
@@ -61,7 +61,7 @@ namespace KvalDarbsCore.Controllers
             return View(competitions.ToList());
         }
 
-        // GET: Competitions/Details/5
+        [HttpGet]
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -69,7 +69,7 @@ namespace KvalDarbsCore.Controllers
                 return NotFound();
             }
 
-            var competition = _context.Competitions
+            var competition = _context.Competitions.Include(m => m.Results).ThenInclude(m => m.Athlete)
                 .FirstOrDefault(m => m.Id == id);
             if (competition == null)
             {
@@ -80,15 +80,12 @@ namespace KvalDarbsCore.Controllers
             return View(competition);
         }
 
-        // GET: Competitions/Create
+        [HttpGet]
         public ActionResult Create()
         {
             return View();
         }
 
-        // POST: Competitions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind("Id,Name,Date,Distance,Type,Location")] Competition competition)
@@ -110,7 +107,7 @@ namespace KvalDarbsCore.Controllers
             return View(competition);
         }
 
-        // GET: Competitions/Edit/5
+        [HttpGet]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -130,9 +127,6 @@ namespace KvalDarbsCore.Controllers
             return View(competition);
         }
 
-        // POST: Competitions/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(int? id, [Bind("Id,Name,Date,Distance,Type,Location")] Competition competition)
@@ -146,7 +140,7 @@ namespace KvalDarbsCore.Controllers
             if (user == null || string.IsNullOrEmpty(user.Id))
                 return Unauthorized();
 
-            if (_context.Competitions.Any(m => m.Id == id && m.AuthorId == user.Id))
+            if (!_context.Competitions.Any(m => m.Id == id && m.AuthorId == user.Id))
             {
                 return NotFound();
             }
@@ -155,6 +149,7 @@ namespace KvalDarbsCore.Controllers
             {
                 try
                 {
+                    competition.AuthorId = user.Id;
                     _context.Update(competition);
                     _context.SaveChanges();
                 }
@@ -174,7 +169,7 @@ namespace KvalDarbsCore.Controllers
             return View(competition);
         }
 
-        // GET: Competitions/Delete/5
+        [HttpGet]
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Major Code Smell", "S4144:Methods should not have identical implementations", Justification = "Returns different view from Details.")]
         public ActionResult Delete(int? id)
         {
@@ -197,7 +192,6 @@ namespace KvalDarbsCore.Controllers
             return View(competition);
         }
 
-        // POST: Competitions/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int? id)
@@ -211,11 +205,18 @@ namespace KvalDarbsCore.Controllers
             if (competition == null)
                 return NotFound();
 
+            if (competition.Results.Any())
+            {
+                ViewData["ResultsExist"] = "Sacensībām ir saistīti rezultāti.";
+                return View("Delete", competition);
+            }
+
             _context.Competitions.Remove(competition);
             _context.SaveChanges();
             return RedirectToAction(nameof(Index));
         }
 
+        [HttpGet]
         public ActionResult CreateResult(int? id)
         {
             ViewData["AthleteId"] = _context.GetActiveUser(this.HttpContext)?.Id;
@@ -223,9 +224,6 @@ namespace KvalDarbsCore.Controllers
             return View();
         }
 
-        // POST: Competitions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult CreateResult([Bind("AthleteId,CompetitionId,Hours,Minutes,Seconds")] Result result)
@@ -243,13 +241,134 @@ namespace KvalDarbsCore.Controllers
             if (ModelState.IsValid)
             {
                 result.Athlete = user;
+                result.AthleteId = user.Id;
                 _context.Add(result);
                 _context.SaveChanges();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Details), new { id = result.CompetitionId });
             }
 
             ViewData["AthleteId"] = user.Id;
             ViewData["CompetitionId"] = result.CompetitionId;
+            return View(result);
+        }
+
+        [HttpGet]
+        public ActionResult EditResult(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = _context.GetActiveUser(this.HttpContext);
+            if (user == null || string.IsNullOrEmpty(user.Id))
+                return Unauthorized();
+
+            var result = _context.Results.FirstOrDefault(m => m.Id == id && m.AthleteId == user.Id);
+            if (result == null)
+            {
+                return NotFound();
+            }
+            return View(result);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditResult(int? id, [Bind("Id,AthleteId,CompetitionId,Hours,Minutes,Seconds")] Result result)
+        {
+            if (id != result.Id)
+            {
+                return NotFound();
+            }
+
+            var user = _context.GetActiveUser(this.HttpContext);
+            if (user == null || string.IsNullOrEmpty(user.Id))
+                return Unauthorized();
+
+            if (!_context.Results.Any(m => m.Id == id && m.AthleteId == user.Id))
+            {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    _context.Update(result);
+                    _context.SaveChanges();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!CompetitionExists(result.CompetitionId))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return RedirectToAction(nameof(Details), new { id = result.CompetitionId });
+            }
+            return View(result);
+        }
+
+        [HttpGet]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Major Code Smell", "S4144:Methods should not have identical implementations", Justification = "Returns different view from Details.")]
+        public ActionResult DeleteResult(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var user = _context.GetActiveUser(this.HttpContext);
+            if (user == null || string.IsNullOrEmpty(user.Id))
+                return Unauthorized();
+
+            var result = _context.Results.Include(m => m.Athlete)
+                .FirstOrDefault(m => m.Id == id && m.AthleteId == user.Id);
+            if (result == null)
+            {
+                return NotFound();
+            }
+
+            return View(result);
+        }
+
+        [HttpPost, ActionName("DeleteResult")]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteResultConfirmed(int? id)
+        {
+            var user = _context.GetActiveUser(this.HttpContext);
+            if (user == null || string.IsNullOrEmpty(user.Id))
+                return Unauthorized();
+
+            var result = _context.Results.FirstOrDefault(m => m.Id == id && m.AthleteId == user.Id);
+
+            if (result == null)
+                return NotFound();
+
+            _context.Results.Remove(result);
+            _context.SaveChanges();
+            return RedirectToAction(nameof(Details), new { id = result.CompetitionId});
+        }
+
+        [HttpGet]
+        public ActionResult DetailsResult(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var result = _context.Results.Include(m => m.Athlete)
+                .FirstOrDefault(m => m.Id == id);
+            if (result == null)
+            {
+                return NotFound();
+            }
+            ViewData["ActiveUserId"] = _context.GetActiveUser(this.HttpContext)?.Id;
             return View(result);
         }
